requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.13.1
joblib==1.4.2
pyyaml==6.0.1
streamlit==1.36.0
plotly==5.22.0
jupyter==1.0.0
ipykernel==6.29.5
